Protocol v2 stage B: wire layer (/tap, manifest/bundle fetch, uint64 ids)
- net_rest: rest_post_tap (v2 action report with outcome parse),
  rest_get_manifest, rest_get_bundle; rest_frame_out_t carries the
  optional "manifest" block (the v2-server signal); status body adds
  "proto": {"v": 2}, overlay.max_glyphs 64, and tz; status response
  parses config.tz (persisted), local_hh/local_mm, and the raw "sync"
  digest-triple passthrough.
- Wake-event ids widened to uint64 end to end (button_report,
  touch_queue, deck_run, net_rest, main). Seeded at power-on with a
  52-bit random start so ids stay <= 2^53: they travel as JSON numbers
  and the cJSON double round-trip must stay exact.

Host suites: proto2 77, overlay 86, deck 42 -- all green.

+++ src/button_report.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-void button_report_set(button_report_t *report, const char *name, uint32_t event_id)
+void button_report_set(button_report_t *report, const char *name, uint64_t event_id)
 {
     if (name && name[0]) snprintf(report->name, sizeof report->name, "%s", name);
     else                report->name[0] = '\0';
@@ -21,8 +21,9 @@ int button_report_append_frame_query(const button_report_t *report,
         return offset;
 
     int written = snprintf(url + offset, url_size - (size_t)offset,
-                           "?button=%s&button_event_id=%u",
-                           report->name, (unsigned)report->event_id);
+                           "?button=%s&button_event_id=%llu",
+                           report->name,
+                           (unsigned long long)report->event_id);
     if (written < 0) return offset;
 
     size_t next = (size_t)offset + (size_t)written;
