+++ src/deck_run.h
@@ -31,7 +31,7 @@ void deck_boot(void);
  * bringing the radio up; true = a later press had no local link, its report
  * has been armed via rest_set_button(), continue into the network cycle.
  * event_seq is main's wake-event counter (shared with touch). */
-bool deck_try_button(const char *button, uint32_t *event_seq, bool *fallthrough);
+bool deck_try_button(const char *button, uint64_t *event_seq, bool *fallthrough);
 
 /* Local nav for a captured touch tap: hit-tests (x, y) -- the stroke end --
  * against the current page's zones. True = painted from SD; the caller must
